backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
pillow
transformers
torch
torchvision
torchaudio
